upgrade to v2.1, remove Test filter, split into WebFilter and SecurityFilter
