+++ pom.xml
@@ -3,7 +3,7 @@
 
     <groupId>org.pac4j.demo</groupId>
     <artifactId>j2e-pac4j-cdi-demo</artifactId>
-    <version>2.0.0-SNAPSHOT</version>
+    <version>2.1.0-SNAPSHOT</version>
     <packaging>war</packaging>
 
     <name>j2e-cdi-pac4j demo</name>
@@ -32,7 +32,7 @@
         <payara-micro-maven-plugin.version>0.0.1</payara-micro-maven-plugin.version>
         <!-- Dependency versions -->
         <dependency.pac4j.version>2.0.0</dependency.pac4j.version>
-        <dependency.j2e-pac4j.version>2.0.0</dependency.j2e-pac4j.version>
+        <dependency.j2e-pac4j.version>2.1.0-SNAPSHOT</dependency.j2e-pac4j.version>
         <dependency.javaee.version>7.0</dependency.javaee.version>
         <dependency.logback.version>1.1.3</dependency.logback.version>
         <dependency.payara-embedded-all.version>4.1.1.171.1</dependency.payara-embedded-all.version>
